Moved single unit order trait.
