+++ src/main/scala/org/economicsl/auctions/singleunit/SingleUnit.scala
@@ -13,9 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions
+package org.economicsl.auctions.singleunit
 
 import org.economicsl.auctions.multiunit.SinglePricePoint
+import org.economicsl.auctions.{Order, Quantity, Tradable}
 
 
 /** Mixin trait that restricts the quantity of an `Order with SinglePricePoint` to a single unit of a `Tradable`. */
